requirements.txt
absl-py==1.4.0
anyio==4.8.0
apache-beam==2.46.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astunparse==1.6.3
attrs==21.4.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
cachetools==4.2.4
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==7.1.2
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
crcmod==1.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
docker==4.4.4
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.2.2
fastavro==1.10.0
fasteners==0.19
fastjsonschema==2.21.1
flask==2.0.3
flatbuffers==25.1.24
fqdn==1.5.1
gast==0.4.0
google-api-core==1.32.0
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.17.1
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.16.2
google-cloud-bigtable==1.7.3
google-cloud-core==2.4.1
google-cloud-datastore==1.15.5
google-cloud-dlp==3.9.2
google-cloud-language==1.3.2
google-cloud-pubsub==2.13.11
google-cloud-pubsublite==1.6.0
google-cloud-recommendations-ai==0.7.1
google-cloud-resource-manager==1.6.3
google-cloud-spanner==3.26.0
google-cloud-storage==2.11.0
google-cloud-videointelligence==1.16.3
google-cloud-vision==3.1.4
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.12.7
grpcio==1.70.0
grpcio-status==1.48.2
h5py==3.12.1
hdfs==2.7.3
httplib2==0.21.0
idna==3.10
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.8.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter==1.1.1
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server==2.10.0
jupyter-server-terminals==0.5.3
jupyterlab==3.5.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.16.3
jupyterlab-widgets==1.1.11
keras==2.10.0
keras-preprocessing==1.1.2
keras-tuner==1.4.7
kt-legacy==1.0.5
kubernetes==12.0.1
libclang==18.1.1
markdown==3.7
markupsafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.1
ml-metadata==1.11.0
ml-pipelines-sdk==1.11.0
nbclassic==1.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==6.5.7
notebook-shim==0.2.4
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
opt-einsum==3.4.0
orjson==3.10.15
overrides==6.5.0
packaging==20.9
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pip==25.0
platformdirs==4.3.6
portpicker==1.6.0
prometheus-client==0.21.1
prompt-toolkit==3.0.50
proto-plus==1.26.0
protobuf==3.19.6
psutil==6.1.1
pyarrow==6.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydot==1.4.2
pyfarmhash==0.3.2
pygments==2.19.1
pymongo==3.13.0
pyparsing==3.2.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
pywin32==227
pywinpty==2.0.14
pyyaml==5.4.1
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
send2trash==1.8.3
setuptools==70.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-data-validation==1.11.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.11.0
tensorflow-model-analysis==0.42.0
tensorflow-serving-api==2.10.1
tensorflow-transform==1.11.0
termcolor==2.5.0
terminado==0.18.1
tfx==1.11.0
tfx-bsl==1.11.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
uri-template==1.3.0
uritemplate==3.0.1
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==3.6.10
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
